Access Control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use DB;
 
 class PostsController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -84,7 +94,11 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
-        //return view('posts.edit')->with('post',$post);
+        
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page!');
+        }
         return view('posts.edit')->with('post',$post);
 
     }
@@ -123,6 +137,11 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+        
+        //Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page!');
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
